test(sqlite3): retire legacy sqlite3-adapter test, move trails-only coverage
The pre-RFC-0026 packages/activerecord/src/adapters/sqlite3-adapter.test.ts
was a bespoke trails file: none of its test names match Rails.

- raw SQL execution / transactions / Migration integration: retired as
  redundant with database-statements, transactions and migration suites.
- pragmas option: the application/override cases are covered by the
  relocated Rails-named pragma tests; only the trails-only `pragmas`
  constructor-option validation is kept.
- lookupCastType and _isMemoryFilename move verbatim to a
  .trails.test.ts sibling under adapters/sqlite3/ per the RFC-0026 layout.
