+++ eslint/no-explicit-any-test-exclude.json
@@ -45,10 +45,10 @@
   "packages/activerecord/src/adapters/postgresql/uuid.test.ts",
   "packages/activerecord/src/adapters/postgresql/virtual-column.test.ts",
   "packages/activerecord/src/adapters/postgresql/xml.test.ts",
-  "packages/activerecord/src/adapters/sqlite3-adapter.test.ts",
   "packages/activerecord/src/adapters/sqlite3/copy-table.test.ts",
   "packages/activerecord/src/adapters/sqlite3/json.test.ts",
   "packages/activerecord/src/adapters/sqlite3/sqlite3-adapter.test.ts",
+  "packages/activerecord/src/adapters/sqlite3/sqlite3-adapter.trails.test.ts",
   "packages/activerecord/src/adapters/sqlite3/transaction.test.ts",
   "packages/activerecord/src/aggregations.test.ts",
   "packages/activerecord/src/associations.test.ts",
